feat: add cheque claim
