+++ devtools/config/config.toml
@@ -12,6 +12,8 @@ network_type = "ckb_testnet"
 # Fill cellbase maturity in the same as the ckb genesis block.
 cellbase_maturity = 4
 
+cheque_since = 6
+
 log_path = "./free-space/mercury.log"
 
 ckb_uri = "http://127.0.0.1:8114"
